Move math_wrap() to function_helpers.
