+++ src/pycel/lib/function_helpers.py
@@ -1,6 +1,13 @@
 import functools
 
-from pycel.excelutil import ERROR_CODES, math_wrap
+from pycel.excelutil import (
+    coerce_to_number,
+    EMPTY,
+    ERROR_CODES,
+    is_number,
+    NUM_ERROR,
+    VALUE_ERROR,
+)
 
 cse_array_functions = dict(
     # function_parameter which takes cse array input
@@ -58,6 +65,28 @@ def error_string_wrapper(f, err_str_args=None):
     return wrapper
 
 
+def math_wrapper(f):
+    """wrapper for functions that take numbers to handle errors"""
+
+    @functools.wraps(f)
+    def wrapper(*args):
+        # this is a bit of a ::HACK:: to quickly address the most common cases
+        # for reasonable math function parameters
+        for arg in args:
+            if arg in ERROR_CODES:
+                return arg
+        if not (is_number(args[0]) or args[0] in (None, EMPTY)):
+            return VALUE_ERROR
+        try:
+            return f(*(0 if a in (None, EMPTY)
+                       else coerce_to_number(a) for a in args))
+        except ValueError as exc:
+            if "math domain error" in str(exc):
+                return NUM_ERROR
+            raise  # pragma: no cover
+    return wrapper
+
+
 def load_functions(names, name_space, modules):
     # load desired functions into namespace from modules
     not_found = set()
@@ -71,7 +100,7 @@ def load_functions(names, name_space, modules):
                 not_found.add(name)
             else:
                 if module.__name__ == 'math':
-                    func = math_wrap(func)
+                    func = math_wrapper(func)
                     func = cse_array_wrapper(func, 0)
 
                 else:
